requirements.txt
numpy
pandas
opencv-python
tensorflow
deepface
gtts
playsound
SpeechRecognition
sounddevice
wavio
pyobjc==9.0.1
pyttsx3
Pillow
tk